refactor(cmd): разнос main.go по тематическим файлам — чистые перемещения
main.go (2522 строки) содержал вперемешку проводку, ~20 типов-адаптеров,
сервис-менеджмент и cleanup-режим. Первый шаг реструктуризации — дословный
перенос всего, кроме func main(), по тематическим файлам пакета main
(по существующей конвенции <тема>_adapters.go):

- service.go — --service start/stop/restart/status, PID-файл, эндпоинт
- cleanup.go — --cleanup режим (runCleanup) + configSaver
- sysenv.go — detectArch, GC-лимиты, uptime, CA-серты, PATH окружения
- boot.go — logStartup, describeListenInterfaces, populateWANModel
- deviceproxy_adapters.go / routing_adapters.go / monitoring_adapters.go /
  singbox_adapters.go / downloader_adapters.go — адаптеры по подсистемам

Поведение не меняется: только перемещения кода и разнос import-блоков.
main.go: 2522 → 1546 строк (остался func main — извлечение фаз следующим
коммитом).
