+++ cmd/awg-manager/main.go
@@ -2,17 +2,12 @@ package main
 
 import (
 	"context"
-	"encoding/json"
-	"errors"
 	"flag"
 	"fmt"
 	"net/http"
 	"os"
-	"os/exec"
 	"os/signal"
 	"path/filepath"
-	"sort"
-	"strconv"
 	"strings"
 	"sync"
 	"sync/atomic"
@@ -21,13 +16,11 @@ import (
 
 	"log/slog"
 	"runtime"
-	"runtime/debug"
 
 	"github.com/hoaxisr/awg-manager/frontend"
 	"github.com/hoaxisr/awg-manager/internal/accesspolicy"
 	"github.com/hoaxisr/awg-manager/internal/api"
 	"github.com/hoaxisr/awg-manager/internal/auth"
-	"github.com/hoaxisr/awg-manager/internal/cleanup"
 	"github.com/hoaxisr/awg-manager/internal/clientroute"
 	"github.com/hoaxisr/awg-manager/internal/connectivity"
 	"github.com/hoaxisr/awg-manager/internal/deviceproxy"
@@ -64,7 +57,6 @@ import (
 	"github.com/hoaxisr/awg-manager/internal/sys/env"
 	"github.com/hoaxisr/awg-manager/internal/sys/kmod"
 	"github.com/hoaxisr/awg-manager/internal/sys/ndmsinfo"
-	"github.com/hoaxisr/awg-manager/internal/sys/netif"
 	"github.com/hoaxisr/awg-manager/internal/sys/osdetect"
 	"github.com/hoaxisr/awg-manager/internal/terminal"
 	"github.com/hoaxisr/awg-manager/internal/testing"
@@ -83,204 +75,10 @@ import (
 	"github.com/hoaxisr/awg-manager/internal/updater"
 )
 
-const (
-	defaultDataDir = "/opt/etc/awg-manager"
-	// pidFile lives on the system tmpfs (cleared on every boot) so an
-	// unclean reboot can never leave a stale PID pointing at whatever
-	// process eventually inherits that PID slot on the next uptime.
-	// /var/run is FHS-canonical and always tmpfs on Keenetic; /opt/var/run
-	// is Entware-persistent storage and was the source of the stale-PID
-	// startup-block bug.
-	pidFile = "/var/run/awg-manager.pid"
-	// legacyPidFile is the pre-move location; one-shot cleanup on startup
-	// removes it after an upgrade so the old file does not linger.
-	legacyPidFile = "/opt/var/run/awg-manager.pid"
-	// serviceStderrLog captures the daemon's stderr under --service start:
-	// panic traces and early-boot warnings that predate the app logger.
-	serviceStderrLog = "/opt/tmp/awg-manager-stderr.log"
-)
+const defaultDataDir = "/opt/etc/awg-manager"
 
 // version is set via ldflags at build time
 var version = "dev"
-
-// buildArch is set via ldflags at build time to one of the awg-manager
-// architecture keys: "mipsel-3.4" | "mips-3.4" | "aarch64-3.10".
-// Empty when running `go run` / `go build ./cmd/awg-manager` directly —
-// detectArch() falls back to runtime.GOARCH-based mapping.
-var buildArch string
-
-// detectArch returns the awg-manager arch key for installer.EmbeddedBinaries.
-// Prefers the build-time -X main.buildArch override; falls back to
-// runtime.GOARCH for dev builds.
-func detectArch() string {
-	if buildArch != "" {
-		return buildArch
-	}
-	switch runtime.GOARCH {
-	case "mipsle":
-		return "mipsel-3.4"
-	case "mips":
-		return "mips-3.4"
-	case "arm64":
-		return "aarch64-3.10"
-	}
-	return ""
-}
-
-// applyGoMemoryLimits tunes THIS process's GC via runtime/debug. Two things
-// deliberately do NOT happen here:
-//   - no os.Setenv: GOGC/GOMEMLIMIT env vars are read once at runtime init,
-//     so setting them from main() is a no-op for the current process — and
-//     the mutated env would be inherited by the spawned sing-box, overriding
-//     its own deliberate defaults (GOGC=75/GOMEMLIMIT=128MiB, see
-//     internal/singbox/process.go) with the manager's much tighter limits.
-//   - explicit env vars still win: a user-provided GOGC/GOMEMLIMIT was
-//     already applied by the runtime at startup, so we skip that knob.
-func applyGoMemoryLimits(disableMemorySaving bool) {
-	for _, entry := range osdetect.GetGCEnv(disableMemorySaving) {
-		parts := strings.SplitN(entry, "=", 2)
-		if len(parts) != 2 || os.Getenv(parts[0]) != "" {
-			continue
-		}
-		switch parts[0] {
-		case "GOGC":
-			if pct, err := strconv.Atoi(parts[1]); err == nil {
-				debug.SetGCPercent(pct)
-			}
-		case "GOMEMLIMIT":
-			if limit, err := parseByteSize(parts[1]); err == nil {
-				debug.SetMemoryLimit(limit)
-			}
-		}
-	}
-}
-
-// parseByteSize parses the GOMEMLIMIT syntax subset osdetect emits ("16MiB").
-func parseByteSize(s string) (int64, error) {
-	n := strings.TrimSuffix(s, "MiB")
-	if n == s {
-		return 0, fmt.Errorf("unsupported byte size %q", s)
-	}
-	mib, err := strconv.ParseInt(strings.TrimSpace(n), 10, 64)
-	if err != nil {
-		return 0, err
-	}
-	return mib << 20, nil
-}
-
-// deviceproxySubscriptionOutboundsAdapter adapts *subscription.Service to
-// the deviceproxy.SubscriptionOutboundsCatalog interface, exposing enabled
-// subscription selector/urltest outbounds as device-proxy targets without
-// creating a direct dependency from the subscription package on deviceproxy.
-type deviceproxySubscriptionOutboundsAdapter struct {
-	src *subscription.Service
-}
-
-func (a *deviceproxySubscriptionOutboundsAdapter) ListDeviceProxyOutbounds() []deviceproxy.SubscriptionOutboundInfo {
-	if a == nil || a.src == nil {
-		return nil
-	}
-	subs := a.src.List()
-	out := make([]deviceproxy.SubscriptionOutboundInfo, 0, len(subs))
-
-	for _, sub := range subs {
-		if !sub.Enabled || sub.SelectorTag == "" || len(sub.MemberTags) == 0 {
-			continue
-		}
-		label := strings.TrimSpace(sub.Label)
-		if label == "" {
-			label = sub.ID
-		}
-		active := sub.ActiveMember
-		if active == "" && len(sub.MemberTags) > 0 {
-			active = sub.MemberTags[0]
-		}
-		detail := active
-		for _, m := range sub.Members {
-			if m.Tag != active {
-				continue
-			}
-			parts := []string{}
-			if strings.TrimSpace(m.Label) != "" {
-				parts = append(parts, strings.TrimSpace(m.Label))
-			}
-			if m.Protocol != "" {
-				parts = append(parts, strings.ToUpper(m.Protocol))
-			}
-			if m.Server != "" {
-				parts = append(parts, fmt.Sprintf("%s:%d", m.Server, m.Port))
-			}
-			if len(parts) > 0 {
-				detail = strings.Join(parts, " · ")
-			}
-			break
-		}
-		out = append(out, deviceproxy.SubscriptionOutboundInfo{
-			Tag:    sub.SelectorTag,
-			Label:  label,
-			Detail: detail,
-		})
-	}
-	sort.Slice(out, func(i, j int) bool {
-		return out[i].Label < out[j].Label
-	})
-	return out
-}
-
-// deviceproxyRouterOutboundsAdapter adapts *router.ServiceImpl to the
-// deviceproxy.RouterOutboundsCatalog interface, exposing router-defined
-// outbounds (20-router.json) as device-proxy targets. Only Source=="router"
-// entries are returned (subscription composites are surfaced separately via
-// SubscriptionOutboundsCatalog). Directs that stripAutoManagedDirect removes
-// from the effective config are hidden — they are not selectable.
-type deviceproxyRouterOutboundsAdapter struct {
-	src *router.ServiceImpl
-}
-
-func (a *deviceproxyRouterOutboundsAdapter) ListDeviceProxyRouterOutbounds() []deviceproxy.RouterOutboundInfo {
-	if a == nil || a.src == nil {
-		return nil
-	}
-	views, err := a.src.ListCompositeOutbounds(context.Background())
-	if err != nil {
-		return nil
-	}
-	out := make([]deviceproxy.RouterOutboundInfo, 0, len(views))
-	for _, v := range views {
-		if v.Source != "router" {
-			continue
-		}
-		o := v.Outbound
-		if o.Type == "direct" && o.BindInterface != "" && router.IsStrippedDirectBind(o.BindInterface) {
-			continue // не попадёт в эффективный конфиг → невыбираемо
-		}
-		detail := ""
-		if o.Type == "direct" {
-			if o.BindInterface != "" {
-				detail = "direct · " + o.BindInterface
-			} else {
-				detail = "direct"
-			}
-		} else {
-			detail = fmt.Sprintf("%s · %d", o.Type, len(o.Outbounds))
-		}
-		out = append(out, deviceproxy.RouterOutboundInfo{
-			Tag:    o.Tag,
-			Label:  o.Tag,
-			Detail: detail,
-			// Определение композита нужно device-proxy для graceful-
-			// деградации: когда слот 20 припаркован (движок выключен),
-			// селектор слота 30 подставляет default-член композита
-			// вместо висячей ссылки на его тег (issue #465).
-			DefaultMember: o.Default,
-			Members:       append([]string(nil), o.Outbounds...),
-		})
-	}
-	sort.Slice(out, func(i, j int) bool {
-		return out[i].Label < out[j].Label
-	})
-	return out
-}
 
 func main() {
 	dataDir := flag.String("data-dir", defaultDataDir, "Data directory path")
@@ -1747,776 +1545,4 @@ func main() {
 		fmt.Fprintf(os.Stderr, "Server error: %v\n", err)
 		os.Exit(1)
 	}
-}
-
-// configSaver adapts *ndmscommand.SaveCoordinator to the cleanup.ConfigSaver
-// interface (Save(ctx) error). Flush forces the debounced NDMS save to run
-// synchronously, matching the contract expected by the cleanup service.
-type configSaver struct {
-	sc *ndmscommand.SaveCoordinator
-}
-
-func (c configSaver) Save(ctx context.Context) error {
-	return c.sc.Flush(ctx)
-}
-
-// tunnelProviderAdapter adapts service.Service to routing.TunnelProvider.
-type tunnelProviderAdapter struct {
-	svc   service.Service
-	store *storage.AWGTunnelStore
-}
-
-func (a *tunnelProviderAdapter) ListTunnels(ctx context.Context) ([]routing.TunnelWithStatus, error) {
-	tunnels, err := a.svc.List(ctx)
-	if err != nil {
-		return nil, err
-	}
-	result := make([]routing.TunnelWithStatus, len(tunnels))
-	for i, t := range tunnels {
-		entry := routing.TunnelWithStatus{
-			ID:      t.ID,
-			Name:    t.Name,
-			Backend: t.Backend,
-			State:   t.State,
-		}
-		// NativeWG tunnels need NWGIndex from storage.
-		if t.Backend == "nativewg" {
-			if stored, err := a.store.Get(t.ID); err == nil {
-				entry.NWGIndex = stored.NWGIndex
-			}
-		}
-		result[i] = entry
-	}
-	return result, nil
-}
-
-func (a *tunnelProviderAdapter) GetState(ctx context.Context, tunnelID string) tunnel.StateInfo {
-	return a.svc.GetState(ctx, tunnelID)
-}
-
-func (a *tunnelProviderAdapter) WANModel() *wan.Model {
-	return a.svc.WANModel()
-}
-
-// dnsRouteCountAdapter adapts dnsroute.Store to dnscheck.DnsRouteProvider.
-type dnsRouteCountAdapter struct {
-	store *dnsroute.Store
-}
-
-func (a *dnsRouteCountAdapter) ListEnabledCount(_ context.Context) (int, int) {
-	data := a.store.GetCached()
-	if data == nil {
-		return 0, 0
-	}
-	total := len(data.Lists)
-	enabled := 0
-	for _, l := range data.Lists {
-		if l.Enabled {
-			enabled++
-		}
-	}
-	return total, enabled
-}
-
-// runningTunnelAdapter adapts service.Service to dnscheck.TunnelStateProvider.
-type runningTunnelAdapter struct {
-	svc service.Service
-}
-
-func (a *runningTunnelAdapter) RunningTunnelNames(ctx context.Context) []string {
-	list, err := a.svc.List(ctx)
-	if err != nil {
-		return nil
-	}
-	var names []string
-	for _, t := range list {
-		if t.State == tunnel.StateRunning {
-			names = append(names, t.Name)
-		}
-	}
-	return names
-}
-
-// storeAdapter adapts storage.AWGTunnelStore to routing.StoreClient.
-type storeAdapter struct {
-	store *storage.AWGTunnelStore
-}
-
-func (a *storeAdapter) Get(id string) (routing.StoreEntry, error) {
-	t, err := a.store.Get(id)
-	if err != nil {
-		return routing.StoreEntry{}, err
-	}
-	return routing.StoreEntry{Backend: t.Backend, NWGIndex: t.NWGIndex}, nil
-}
-
-func (a *storeAdapter) Exists(id string) bool {
-	return a.store.Exists(id)
-}
-
-// logStartup logs system startup information.
-// describeListenInterfaces renders the configured bind scope for logs.
-func describeListenInterfaces(ifaces []string) string {
-	if len(ifaces) == 0 {
-		return "all (0.0.0.0)"
-	}
-	return strings.Join(ifaces, ", ")
-}
-
-func logStartup(appLog *logging.ScopedLogger, version, osVersion, listenAddr string, settings *storage.Settings) {
-	appLog.Info("startup", "", fmt.Sprintf("AWG Manager v%s started", version))
-	appLog.Info("startup", "", fmt.Sprintf("Keenetic OS: %s", osVersion))
-	appLog.Info("startup", "", fmt.Sprintf("Listening on %s", listenAddr))
-
-	// Log feature status
-	if settings.PingCheck.Enabled {
-		appLog.Info("startup", "", "Ping Check: enabled")
-	}
-	if settings.Logging.Enabled {
-		appLog.Info("startup", "", "Logging: enabled")
-	}
-}
-
-// populateWANModel queries NDMS for current WAN interfaces and fills the
-// unified WAN model so that AnyUp() works before any WAN hooks fire.
-func populateWANModel(ctx context.Context, queries *ndmsquery.Queries, model *wan.Model, appLog *logging.ScopedLogger) {
-	interfaces, err := queries.Interfaces.ListWAN(ctx)
-	if err != nil {
-		appLog.Warn("populate-wan", "", "failed to get WAN interfaces: "+err.Error())
-		return
-	}
-	model.Populate(interfaces)
-	ifaceList := make([]string, 0, len(interfaces))
-	for _, iface := range interfaces {
-		ifaceList = append(ifaceList, fmt.Sprintf("%s(up=%t)", iface.Name, iface.Up))
-	}
-	appLog.Info("populate-wan", "", fmt.Sprintf("WAN model populated, count=%d ifaces=[%s]", len(interfaces), strings.Join(ifaceList, " ")))
-}
-
-// getUptime reads system uptime in seconds from /proc/uptime.
-// Returns 0 on error (treated as non-boot scenario).
-func getUptime() float64 {
-	data, err := os.ReadFile("/proc/uptime")
-	if err != nil {
-		return 0
-	}
-	fields := strings.Fields(string(data))
-	if len(fields) == 0 {
-		return 0
-	}
-	uptime, err := strconv.ParseFloat(fields[0], 64)
-	if err != nil {
-		return 0
-	}
-	return uptime
-}
-
-// runService handles --service flag: start/stop/restart/status.
-// This replaces the shell logic that was previously in S99awg-manager.
-func runService(action, dataDir string) {
-	switch action {
-	case "start":
-		serviceStart(dataDir)
-	case "stop":
-		serviceStop()
-	case "restart":
-		serviceStop()
-		time.Sleep(time.Second)
-		serviceStart(dataDir)
-	case "status":
-		serviceStatus(dataDir)
-	default:
-		fmt.Fprintf(os.Stderr, "Unknown service action: %s\nUsage: --service {start|stop|restart|status}\n", action)
-		os.Exit(1)
-	}
-}
-
-// serviceStart starts the daemon as a background process with PID file management.
-func serviceStart(dataDir string) {
-	// Check if already running
-	if pid, running := readPIDFile(); running {
-		fmt.Printf("AWG Manager already running (PID %d)\n", pid)
-		return
-	}
-
-	fmt.Println("Starting AWG Manager...")
-
-	// Ensure directories. /var/run is system tmpfs and always exists,
-	// but MkdirAll is idempotent so harmless to call.
-	os.MkdirAll("/var/run", 0755)
-	os.MkdirAll("/opt/var/log", 0755)
-	os.MkdirAll(dataDir, 0755)
-
-	// Resolve executable path
-	executable, err := os.Executable()
-	if err != nil {
-		executable = os.Args[0]
-	}
-
-	// Ensure system binaries and libraries are available for child processes
-	ensureServiceEnv()
-
-	// Start the daemon without --service flag
-	cmd := exec.Command(executable, "-data-dir", dataDir)
-	setServiceSysProcAttr(cmd)
-
-	// O_RDWR, not os.Open (O_RDONLY): writes to a read-only /dev/null fd fail
-	// with EBADF, which used to silently eat every stderr warning and panic
-	// trace of the daemon.
-	devNull, err := os.OpenFile(os.DevNull, os.O_RDWR, 0)
-	if err == nil {
-		cmd.Stdout = devNull
-		cmd.Stderr = devNull
-		cmd.Stdin = devNull
-		defer devNull.Close()
-	}
-	// Keep panic traces and early-boot warnings (bind fallbacks, corrupt
-	// settings recovery) somewhere findable instead of /dev/null.
-	if logf, lerr := os.OpenFile(serviceStderrLog, os.O_WRONLY|os.O_CREATE|os.O_TRUNC, 0644); lerr == nil {
-		cmd.Stderr = logf
-		defer logf.Close()
-	}
-
-	if err := cmd.Start(); err != nil {
-		fmt.Fprintf(os.Stderr, "Failed to start AWG Manager: %v\n", err)
-		os.Exit(1)
-	}
-
-	childPID := cmd.Process.Pid
-
-	// Write PID file
-	_ = os.WriteFile(pidFile, []byte(strconv.Itoa(childPID)+"\n"), 0644)
-
-	// Detach from child — it becomes an orphan re-parented to init
-	cmd.Process.Release()
-
-	// Wait for process to start (up to 5 seconds)
-	for i := 0; i < 5; i++ {
-		time.Sleep(time.Second)
-		if isProcessRunning(childPID) {
-			host, port := getServiceEndpoint(dataDir)
-			fmt.Printf("AWG Manager started: http://%s:%d\n", host, port)
-			return
-		}
-	}
-
-	fmt.Fprintln(os.Stderr, "AWG Manager failed to start")
-	os.Remove(pidFile)
-	os.Exit(1)
-}
-
-// serviceStop stops the running daemon via PID file.
-func serviceStop() {
-	pid, running := readPIDFile()
-	if !running {
-		fmt.Println("AWG Manager stopped")
-		return
-	}
-
-	fmt.Println("Stopping AWG Manager...")
-
-	process, err := os.FindProcess(pid)
-	if err != nil {
-		os.Remove(pidFile)
-		fmt.Println("AWG Manager stopped")
-		return
-	}
-
-	// Send SIGTERM for graceful shutdown
-	_ = process.Signal(syscall.SIGTERM)
-
-	// Wait up to 5 seconds for process to exit
-	for i := 0; i < 5; i++ {
-		time.Sleep(time.Second)
-		if !isProcessRunning(pid) {
-			break
-		}
-	}
-
-	// Force kill if still running
-	if isProcessRunning(pid) {
-		_ = process.Signal(syscall.SIGKILL)
-	}
-
-	os.Remove(pidFile)
-	fmt.Println("AWG Manager stopped")
-}
-
-// serviceStatus checks if the daemon is running and prints its endpoint.
-func serviceStatus(dataDir string) {
-	pid, running := readPIDFile()
-	if !running {
-		fmt.Println("AWG Manager not running")
-		os.Exit(1)
-	}
-
-	host, port := getServiceEndpoint(dataDir)
-	fmt.Printf("AWG Manager running (PID %d): http://%s:%d\n", pid, host, port)
-}
-
-// readPIDFile reads the PID file and checks if the process is alive.
-// Returns the PID and whether the process is running.
-func readPIDFile() (int, bool) {
-	data, err := os.ReadFile(pidFile)
-	if err != nil {
-		return 0, false
-	}
-	pid, err := strconv.Atoi(strings.TrimSpace(string(data)))
-	if err != nil || pid <= 0 {
-		return 0, false
-	}
-	if !isProcessRunning(pid) {
-		// Stale PID file
-		os.Remove(pidFile)
-		return 0, false
-	}
-	return pid, true
-}
-
-// isProcessRunning checks if a process with the given PID is an awg-manager
-// instance. /proc/<pid>/cmdline is the NUL-separated argv. We match on the
-// basename of argv[0] rather than on the whole buffer so an argument that
-// happens to contain "awg-manager" (e.g. "-data-dir /opt/etc/awg-manager")
-// for an unrelated process that inherited the recycled PID does not
-// produce a false positive.
-func isProcessRunning(pid int) bool {
-	data, err := os.ReadFile(fmt.Sprintf("/proc/%d/cmdline", pid))
-	if err != nil {
-		return false
-	}
-	argv0 := string(data)
-	if i := strings.IndexByte(argv0, 0); i >= 0 {
-		argv0 = argv0[:i]
-	}
-	return filepath.Base(argv0) == "awg-manager"
-}
-
-// getServiceEndpoint reads settings to determine the service host:port for display.
-func getServiceEndpoint(dataDir string) (string, int) {
-	port := 2222
-	settingsFile := filepath.Join(dataDir, "settings.json")
-	if data, err := os.ReadFile(settingsFile); err == nil {
-		var s struct {
-			Server struct {
-				Port int `json:"port"`
-			} `json:"server"`
-		}
-		if json.Unmarshal(data, &s) == nil && s.Server.Port > 0 {
-			port = s.Server.Port
-		}
-	}
-
-	// Use br0 (LAN bridge) for display — this is what the user connects from
-	host := netif.FirstIPv4(storage.DefaultInterface)
-	if host == "" {
-		host = "192.168.1.1"
-	}
-
-	return host, port
-}
-
-// ensureCACerts sets SSL_CERT_FILE for entware-based systems (Keenetic) where
-// CA certificates live in /opt/etc/ssl/ instead of standard Linux paths.
-// Without this, Go's crypto/tls fails to verify GitHub (and other) certificates.
-func ensureCACerts() {
-	if os.Getenv("SSL_CERT_FILE") != "" {
-		return
-	}
-	const entwareCert = "/opt/etc/ssl/certs/ca-certificates.crt"
-	if _, err := os.Stat(entwareCert); err == nil {
-		os.Setenv("SSL_CERT_FILE", entwareCert)
-	}
-}
-
-// ensureServiceEnv ensures PATH contains Entware and system directories so
-// child processes can find binaries by name. Entware dirs go FIRST: a bare
-// "ip" must resolve to iproute2 (/opt/sbin/ip), not the firmware busybox
-// applet that lacks `ip rule`/`route show table` features. The guard checks
-// for /opt/bin specifically — an ndm-hook environment can already contain
-// /usr/sbin yet miss the Entware dirs entirely.
-// LD_LIBRARY_PATH is intentionally NOT set: forcing /lib:/usr/lib first
-// poisons Entware binaries (curl/openssl) by making ld.so load incompatible
-// system libraries → SIGSEGV/SIGBUS at runtime.
-func ensureServiceEnv() {
-	path := os.Getenv("PATH")
-	if !strings.Contains(path, "/opt/bin") || !strings.Contains(path, "/usr/sbin") {
-		os.Setenv("PATH", "/opt/bin:/opt/sbin:/bin:/sbin:/usr/bin:/usr/sbin:"+path)
-	}
-}
-
-// runCleanup removes all awg-manager resources and config files.
-// Called during package uninstall (opkg remove).
-func runCleanup(dataDir string) {
-	fmt.Println("awg-manager cleanup: removing all managed resources...")
-
-	settingsStore := storage.NewSettingsStore(dataDir)
-	settingsStore.Load()
-
-	loggingService := logging.NewService(settingsStore)
-	defer loggingService.Stop()
-	bootLog := logging.NewScopedLogger(loggingService, logging.GroupSystem, logging.SubCleanup)
-
-	awgStore := storage.NewAWGTunnelStore(filepath.Join(dataDir, "tunnels"))
-
-	// Build minimal NDMS CQRS layer first — state.Manager consumes
-	// Queries.Interfaces, and ProxyManager / dnsroute / accesspolicy share
-	// the same transport + commands further down.
-	cleanupEventBus := events.NewBus()
-	cleanupNDMSTransport := ndmstransport.New(ndmstransport.NewSemaphore(4))
-	cleanupNDMSQueries := ndmsquery.NewQueries(ndmsquery.Deps{
-		Getter: cleanupNDMSTransport,
-		Logger: nil,
-		IsOS5:  osdetect.Is5,
-	})
-
-	// Init NDMS info (needed for OS detection). Wire ndmsinfo to the
-	// SystemInfoStore, then initialize with retry.
-	if err := ndmsinfo.Init(context.Background(), cleanupNDMSQueries.SystemInfo, 10*time.Second); err != nil {
-		bootLog.Warn("ndms-version", "", err.Error())
-	}
-
-	// Create service components
-	wgClient := wg.New()
-	backendImpl := backend.New(bootLog)
-	stateMgr := state.New(cleanupNDMSQueries.Interfaces, wgClient, backendImpl, nil)
-	firewallMgr := firewall.New(backendImpl.Type() == backend.TypeKernel, osdetect.Is5(), nil)
-
-	// Build NDMS Commands early so the Operator can consume them. HookNotifier
-	// is wired below once the orchestrator exists (see SetHookNotifier call).
-	cleanupNDMSSave := ndmscommand.NewSaveCoordinator(
-		cleanupNDMSTransport,
-		cleanupEventBus,
-		3*time.Second,
-		10*time.Second,
-		env.DurationDefault("AWG_NDMS_SAVE_SETTLE_DELAY", 2*time.Second),
-		cleanupNDMSQueries.RunningConfig,
-	)
-	cleanupNDMSCommands := ndmscommand.NewCommands(ndmscommand.Deps{
-		Poster:  cleanupNDMSTransport,
-		Save:    cleanupNDMSSave,
-		Queries: cleanupNDMSQueries,
-		IsOS5:   osdetect.Is5,
-	})
-
-	operator := ops.NewOperator(cleanupNDMSQueries, cleanupNDMSCommands, wgClient, backendImpl, firewallMgr)
-
-	nwgOp := nwg.NewOperator(cleanupNDMSQueries, cleanupNDMSCommands, cleanupNDMSTransport, nil)
-	tunnelService := service.New(awgStore, nwgOp, operator, stateMgr, wan.NewModel(), nil)
-
-	// Wire orchestrator for lifecycle operations (Delete needs it)
-	cleanupOrch := orchestrator.New(awgStore, operator, nwgOp, stateMgr, wan.NewModel(), nil)
-	tunnelService.SetOrchestrator(cleanupOrch)
-	nwgOp.SetHookNotifier(cleanupOrch)
-	if os5Op, ok := operator.(interface {
-		SetHookNotifier(tunnel.HookNotifier)
-	}); ok {
-		os5Op.SetHookNotifier(cleanupOrch)
-	}
-	// Wire HookNotifier on NDMS Commands now that the orchestrator exists.
-	cleanupNDMSCommands.SetHookNotifier(cleanupOrch)
-
-	// Create auxiliary services
-	dnsStore := dnsroute.NewStore(dataDir)
-	dnsStore.Load()
-	// dnsSvc is constructed later, after cleanup NDMS CQRS layer is built.
-
-	// Client route service for cleanup
-	clientRouteStore := storage.NewClientRouteStore(dataDir)
-	clientRouteSvc := clientroute.New(clientRouteStore, operator, nil, nil)
-
-	// Managed WireGuard server — wired to the cleanup-path NDMS CQRS layer.
-	managedSvc := managed.New(
-		cleanupNDMSTransport,
-		cleanupNDMSSave,
-		cleanupNDMSQueries,
-		cleanupNDMSCommands,
-		settingsStore,
-		slog.Default(),
-		nil,
-	)
-
-	// DNS route service wired to cleanup NDMS CQRS layer (OS5 only — OS4
-	// short-circuits inside reconcile via ErrNotSupportedOnOS4).
-	dnsSvc := dnsroute.NewService(dnsStore, cleanupNDMSQueries, cleanupNDMSCommands, nil, nil)
-
-	singboxOp := singbox.NewOperator(singbox.OperatorDeps{
-		Log:                slog.Default().With("component", "singbox"),
-		Queries:            cleanupNDMSQueries,
-		Commands:           cleanupNDMSCommands,
-		SingboxLogLevel:    settingsStore.GetSingboxLogLevel,
-		IsNDMSProxyEnabled: settingsStore.IsSingboxNDMSProxyEnabled,
-	})
-
-	// Cleanup mode: bootstrap the orchestrator so any subsequent
-	// operator call that goes through ApplyConfig writes the slot
-	// file rather than the legacy in-place tunnels.json. Cleanup
-	// itself only invokes singboxOp.Cleanup, but we keep the wiring
-	// symmetrical to the daemon path so future cleanup steps have it
-	// available.
-	cleanupSingboxConfigDir := singboxOp.ConfigDir()
-	if err := singbox.MigrateDeviceProxyOutOfTunnels(cleanupSingboxConfigDir); err != nil {
-		bootLog.Warn("deviceproxy-migration", "", err.Error())
-	}
-	cleanupSbOrch := singboxorch.New(cleanupSingboxConfigDir, singboxOp.Process())
-	for _, meta := range singboxorch.KnownSlots() {
-		if err := cleanupSbOrch.Register(meta); err != nil {
-			bootLog.Error("singbox-orchestrator", string(meta.Slot), "register failed: "+err.Error())
-		}
-	}
-	if err := cleanupSbOrch.Bootstrap(); err != nil {
-		bootLog.Error("singbox-orchestrator", "bootstrap", err.Error())
-	}
-	singboxOp.SetOrch(cleanupSbOrch)
-
-	accessPolicySvc := accesspolicy.New(cleanupNDMSCommands.Policies, cleanupNDMSCommands.Interfaces, cleanupNDMSQueries, settingsStore, nil, ndmsquery.NewPolicyMarkStore(cleanupNDMSTransport, nil))
-
-	ctx, cancel := context.WithTimeout(context.Background(), 60*time.Second)
-	defer cancel()
-
-	// Single cleanup call — all business logic in CleanupService
-	cleanupSvc := cleanup.New(tunnelService, awgStore, dnsSvc, managedSvc, accessPolicySvc, clientRouteSvc, singboxOp, configSaver{sc: cleanupNDMSSave})
-	if err := cleanupSvc.CleanupAll(ctx); err != nil {
-		fmt.Fprintf(os.Stderr, "Cleanup error: %v\n", err)
-	}
-
-	// Remove all config/runtime files
-	fmt.Println("Cleaning up files...")
-	os.RemoveAll(filepath.Join(dataDir, "tunnels"))
-	files, _ := filepath.Glob(filepath.Join(dataDir, "*.conf"))
-	for _, f := range files {
-		os.Remove(f)
-	}
-	os.Remove(filepath.Join(dataDir, "port"))
-	os.Remove(filepath.Join(dataDir, "dns-routes.json"))
-
-	fmt.Println("Done.")
-}
-
-// monitoringSystemTunnelAdapter adapts systemtunnel.Service to
-// monitoring.SystemTunnelLister (a small typed view of just the fields the
-// monitoring scheduler needs).
-type monitoringSystemTunnelAdapter struct {
-	svc *systemtunnel.ServiceImpl
-}
-
-func (s *monitoringSystemTunnelAdapter) List(ctx context.Context) ([]monitoring.SystemTunnelInfo, error) {
-	if s == nil || s.svc == nil {
-		return nil, nil
-	}
-	list, err := s.svc.List(ctx)
-	if err != nil {
-		return nil, err
-	}
-	out := make([]monitoring.SystemTunnelInfo, 0, len(list))
-	for _, t := range list {
-		// Skip awg-manager's own server interface (tagged with the
-		// ManagedServerDescription prefix "AWGM ..."; see
-		// internal/managed/types.go::ManagedServerDescription). Server-side
-		// WG is not a client tunnel and must not appear in monitoring.
-		if strings.HasPrefix(t.Description, "AWGM") {
-			continue
-		}
-		out = append(out, monitoring.SystemTunnelInfo{
-			ID:            t.ID,
-			InterfaceName: t.InterfaceName,
-			Description:   t.Description,
-			Connected:     t.Connected,
-		})
-	}
-	return out, nil
-}
-
-// orchLifecycleAdapter routes accesspolicy.SetInterfaceUp for managed tunnels
-// to the orchestrator lifecycle (full start/stop incl. NativeWG kmod proxy).
-type orchLifecycleAdapter struct{ orch *orchestrator.Orchestrator }
-
-func (a orchLifecycleAdapter) Start(ctx context.Context, id string) error {
-	err := a.orch.HandleEvent(ctx, orchestrator.Event{Type: orchestrator.EventStart, Tunnel: id})
-	if errors.Is(err, tunnel.ErrAlreadyRunning) {
-		return nil // already running — user's "on" intent fulfilled
-	}
-	return err
-}
-
-func (a orchLifecycleAdapter) Stop(ctx context.Context, id string) error {
-	return a.orch.HandleEvent(ctx, orchestrator.Event{Type: orchestrator.EventStop, Tunnel: id})
-}
-
-// storeManagedTunnelResolver maps an NDMS interface name to a managed tunnel
-// ID by scanning the AWG tunnel store. ok=false for plain system interfaces.
-type storeManagedTunnelResolver struct{ store *storage.AWGTunnelStore }
-
-func (r storeManagedTunnelResolver) ManagedTunnelByNDMSName(_ context.Context, ndmsName string) (string, bool) {
-	tunnels, err := r.store.List()
-	if err != nil {
-		return "", false
-	}
-	for _, t := range tunnels {
-		var nm string
-		if t.Backend == "nativewg" {
-			nm = nwg.NewNWGNames(t.NWGIndex).NDMSName
-		} else {
-			nm = tunnel.NewNames(t.ID).NDMSName
-		}
-		if nm != "" && nm == ndmsName {
-			return t.ID, true
-		}
-	}
-	return "", false
-}
-
-// operatorLifecycle adapts *singbox.Operator to installer.Lifecycle so
-// the installer can stop/start the daemon during migration without the
-// installer package taking a circular dependency on singbox.
-type operatorLifecycle struct {
-	op *singbox.Operator
-}
-
-func (l *operatorLifecycle) Stop(ctx context.Context) error {
-	return l.op.Control(ctx, "stop")
-}
-
-func (l *operatorLifecycle) Start(ctx context.Context) error {
-	return l.op.Control(ctx, "start")
-}
-
-type installerDownloaderAdapter struct {
-	svc *downloader.Service
-}
-
-type dnsRouteDownloaderAdapter struct {
-	svc *downloader.Service
-}
-
-func (a *dnsRouteDownloaderAdapter) ReadAll(
-	ctx context.Context,
-	req dnsroute.SubscriptionDownloadRequest,
-) ([]byte, dnsroute.SubscriptionDownloadMeta, error) {
-	if a == nil || a.svc == nil {
-		return nil, dnsroute.SubscriptionDownloadMeta{}, fmt.Errorf("downloader service is not configured")
-	}
-	body, meta, err := a.svc.ReadAll(ctx, downloader.Request{
-		Purpose:       "dnsroute-subscription",
-		URL:           req.URL,
-		Timeout:       req.Timeout,
-		MaxBodyBytes:  req.MaxBodyBytes,
-		AllowedStatus: req.AllowedStatus,
-	})
-	if err != nil {
-		return nil, dnsroute.SubscriptionDownloadMeta{}, err
-	}
-	return body, dnsroute.SubscriptionDownloadMeta{ContentType: meta.ContentType}, nil
-}
-
-func (a *installerDownloaderAdapter) DownloadFile(ctx context.Context, req installer.DownloadFileRequest) (installer.DownloadFileResult, error) {
-	if a == nil || a.svc == nil {
-		return installer.DownloadFileResult{}, fmt.Errorf("downloader service is not configured")
-	}
-
-	res, err := a.svc.DownloadFile(ctx, downloader.FileRequest{
-		Request: downloader.Request{
-			Purpose: "singbox-binary",
-			URL:     req.URL,
-			Timeout: req.Timeout,
-		},
-		// Intentional: req.DestPath is already "<binary>.tmp" from installer.
-		// We keep temp == dest here, and activation of live binary is done
-		// separately in Installer.Activate().
-		DestPath:     req.DestPath,
-		TempPath:     req.DestPath,
-		MaxFileBytes: req.MaxFileBytes,
-		Mode:         req.Mode,
-		// Intentional: do not atomically move to final binary here.
-		// Installer.Activate() performs chmod + final atomic rename.
-		Atomic:   false,
-		Progress: req.Progress,
-	})
-	if err != nil {
-		return installer.DownloadFileResult{}, err
-	}
-	return installer.DownloadFileResult{
-		Path: res.Path,
-		Size: res.Size,
-	}, nil
-}
-
-// singboxAndSubLister satisfies singbox.tunnelLister by combining the regular
-// sing-box tunnel list with the active outbound tags of enabled subscriptions.
-// This lets DelayChecker probe subscription active members with the same
-// periodic clash latency test it runs for regular sing-box tunnels.
-type singboxAndSubLister struct {
-	op  *singbox.Operator
-	sub *subscription.Service
-}
-
-func (l *singboxAndSubLister) ListTunnels(ctx context.Context) ([]singbox.TunnelInfo, error) {
-	return l.op.ListTunnels(ctx)
-}
-
-func (l *singboxAndSubLister) ListSubActiveTags() []string {
-	return l.sub.ListActiveMemberTags()
-}
-
-// orchValidatorAdapter bridges singbox.Validator (no context) to the
-// singboxorch.DraftValidator interface (with context).
-type orchValidatorAdapter struct {
-	v *singbox.Validator
-}
-
-func (a *orchValidatorAdapter) Validate(ctx context.Context, configDir string) error {
-	return a.v.Validate(configDir)
-}
-
-// subProxySet adapts the subscription store to singbox.SubscriptionProxySet,
-// exposing each subscription's allocated NDMS composite proxy (index/port/label)
-// so the NDMS-proxy migration and orphan cleanup manage them. Сводные группы
-// (#372) включены в тот же набор — иначе orphan cleanup реап их ProxyN.
-type subProxySet struct {
-	store  *subscription.Store
-	groups *subscription.GroupStore
-}
-
-func (a subProxySet) SubscriptionProxies() []singbox.SubscriptionProxy {
-	if a.store == nil {
-		return nil
-	}
-	var out []singbox.SubscriptionProxy
-	for _, sub := range a.store.List() {
-		if sub.ProxyIndex < 0 || sub.ListenPort == 0 {
-			continue
-		}
-		out = append(out, singbox.SubscriptionProxy{
-			Index: sub.ProxyIndex,
-			Port:  int(sub.ListenPort),
-			Label: sub.Label,
-		})
-	}
-	if a.groups != nil {
-		for _, g := range a.groups.List() {
-			if g.ProxyIndex < 0 || g.ListenPort == 0 {
-				continue
-			}
-			out = append(out, singbox.SubscriptionProxy{
-				Index: g.ProxyIndex,
-				Port:  int(g.ListenPort),
-				Label: g.Label,
-			})
-		}
-	}
-	return out
-}
-
-// dnsRewriteOrchAdapter adapts *singboxorch.Orchestrator to the
-// dnsrewrite.Orchestrator interface (which uses plain string so the
-// package stays decoupled from singboxorch.Slot).
-type dnsRewriteOrchAdapter struct {
-	orch *singboxorch.Orchestrator
-}
-
-func (a *dnsRewriteOrchAdapter) Save(slot string, data []byte) error {
-	return a.orch.Save(singboxorch.Slot(slot), data)
-}
-
-func (a *dnsRewriteOrchAdapter) SetEnabled(slot string, on bool) error {
-	return a.orch.SetEnabled(singboxorch.Slot(slot), on)
 }
